setup environment variable for gtags-cscope

+++ gtags.setup.sh
@@ -1,24 +1,32 @@
 #! /bin/bash
 
+mode = ${1:-"full"}
+
 export GTAGSROOT="~/code"
 export GTAGSDBPATH="~/.vim/caches"
 
-if [ -e "${GTAGSROOT}/GTAGS" ];then
-    
-    rm $GTAGSROOT/GTAGS 
+if [ $mode != "env" ];then
+
+    if [ -e "${GTAGSROOT}/GTAGS" ];then
+
+        rm $GTAGSROOT/GTAGS 
+
+    fi
+
+    if [ -e "${GTAGSROOT}/GRTAGS" ];then
+
+        rm $GTAGSROOT/GRTAGS 
+
+    fi
+
+    if [ -e "${GTAGSROOT}/GPATH" ];then
+
+        rm $GTAGSROOT/GPATH
+
+    fi
+
+    gtags $GTAGSDBPATH
 
 fi
 
-if [ -e "${GTAGSROOT}/GRTAGS" ];then
-    
-    rm $GTAGSROOT/GRTAGS 
 
-fi
-
-if [ -e "${GTAGSROOT}/GPATH" ];then
-    
-    rm $GTAGSROOT/GPATH
-
-fi
-
-gtags $GTAGSDBPATH
